+++ CHANGELOG.md
@@ -68,8 +68,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added global options to define the PCB details (`pcb_material`,
   `solder_mask_color`, `silk_screen_color` and `pcb_finish`)
 - Report generation (for design house) (#93)
-- New output to print PCB layers in SVG format.
-- New output to join PDFs. (#156)
+- New outputs:
+  - Print PCB layers in SVG format.
+  - Join PDFs. (#156)
+  - Export PCB in GENCAD format. (#159)
 
 ### Changed
 - Internal BoM: now components with different Tolerance, Voltage, Current